+++ accessories/tv.js
@@ -18,6 +18,8 @@ class TVAccessory extends BroadlinkRMAccessory {
 
     config.offDuration = config.offDuration || 60;
     config.onDuration = config.onDuration || 60;
+    
+    config.subType = config.subType || 'tv';
 
     if (
       config.enableAutoOn === undefined &&