+++ platform.js
@@ -35,6 +35,7 @@ const BroadlinkRMPlatform = class extends HomebridgePlatform {
 
   addAccessories (accessories) {
     const { config, log } = this;
+    const { debug } = config;
 
     this.discoverBroadlinkDevices();
     this.showMessage();
@@ -68,24 +69,29 @@ const BroadlinkRMPlatform = class extends HomebridgePlatform {
       const homeKitAccessory = new classTypes[accessory.type](log, accessory);
 
       if (classTypes[accessory.type] === classTypes.tv) {
+        if(accessory.subType.toLowerCase() === 'stb'){homeKitAccessory.subType = homebridgeRef.hap.Accessory.Categories.TV_SET_TOP_BOX;}
+        if(accessory.subType.toLowerCase() === 'receiver'){homeKitAccessory.subType = homebridgeRef.hap.Accessory.Categories.AUDIO_RECEIVER;}
+        if(accessory.subType.toLowerCase() === 'stick'){homeKitAccessory.subType = homebridgeRef.hap.Accessory.Categories.TV_STREAMING_STICK;}
+        
+        if (debug) log(`\x1b[34m[DEBUG]\x1b[0m Adding Accessory ${accessory.type} (${accessory.subType})`);
         tvs.push(homeKitAccessory);
         return;
       }
 
+      if (debug) log(`\x1b[34m[DEBUG]\x1b[0m Adding Accessory ${accessory.type} (${accessory.subType})`);
       accessories.push(homeKitAccessory);
     });
 
     if (tvs.length > 0) {
-      accessories.push(tvs.shift(0, 1));
       if (tvs.length > 0) {
         const TV = homebridgeRef.hap.Accessory.Categories.TELEVISION;
-        homebridgeRef.publishExternalAccessories(this, tvs.map(tv => createAccessory(tv, tv.name, TV, homebridgeRef)));
+        homebridgeRef.publishExternalAccessories(this, tvs.map(tv => createAccessory(tv, tv.name, TV, homebridgeRef, tv.subType)));
 
         log('');
         log(`**************************************************************************************************************`);
-        log(`You added more than TVs in your configuration!`);
-        log(`Due to a HomeKit limitation you need to add any additional TV to the Home app by using the Add Accessory function.`);
-        log(`There you'll find your additional TVs and you can use the same PIN as you using for this HomeBridge instance.`);
+        log(`You added TVs in your configuration!`);
+        log(`Due to a HomeKit limitation you need to add any TVs to the Home app by using the Add Accessory function.`);
+        log(`There you'll find your TVs and you can use the same PIN as you using for this HomeBridge instance.`);
         log(`**************************************************************************************************************`);
         log('');
       }